last commit for today

+++ src/main/java/com/diploma/maksimov/restcontroller/CarController.java
@@ -2,6 +2,7 @@ package com.diploma.maksimov.restcontroller;
 
 
 import com.diploma.maksimov.dto.Car;
+import com.diploma.maksimov.service.DriverService;
 import com.diploma.maksimov.service.ICarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/rest/boss/employee/{id}/driver")
+@RequestMapping("/rest/boss/employee")
 
 public class CarController {
     private final ICarService carService;
@@ -21,13 +22,19 @@ public class CarController {
         this.carService = carService;
     }
 
-    @PostMapping(value = "/car")
-    public ResponseEntity<Long> create(@RequestBody Car car) {
+    @Autowired
+    private DriverService driverService;
+
+    @PostMapping(value = "/{id}/driver/car")
+    public ResponseEntity<Long> create(@RequestBody Car car, @PathVariable long id) {
+        if (driverService.read(id) != null) {
             Long idBD = carService.create(car);
             return new ResponseEntity<>(idBD, HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @GetMapping(value = "/car")
+    @GetMapping(value = "/driver/car")
     public ResponseEntity<List<Car>> readAll() {
         final List<Car> cars = carService.readAll();
         if (cars != null && !cars.isEmpty()) {
@@ -36,30 +43,39 @@ public class CarController {
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @GetMapping(value = "/car/{id}")
-    public ResponseEntity<Car> read(@PathVariable long id) {
-        final Car car = carService.read(id);
-        if (car != null) {
-            return new ResponseEntity<>(car, HttpStatus.OK);
+    @GetMapping(value = "/{employeeId}/driver/car/{id}")
+    public ResponseEntity<Car> read(@PathVariable long id,@PathVariable long employeeId) {
+        if (driverService.read(employeeId) != null) {
+            final Car car = carService.read(id);
+            if (car != null) {
+                return new ResponseEntity<>(car, HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @PutMapping(value = "/car/{id}")
-    public ResponseEntity<?> update(@PathVariable long id, @RequestBody Car car) {
-        final boolean updated = carService.update(car, id);
-        if (updated) {
-            return new ResponseEntity<>(HttpStatus.OK);
+    @PutMapping(value = "/{employeeId}/driver/car/{id}")
+    public ResponseEntity<?> update(@PathVariable long id, @RequestBody Car car,@PathVariable long employeeId) {
+        if (driverService.read(employeeId) != null) {
+            final boolean updated = carService.update(car, id);
+            if (updated) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @DeleteMapping(value = "/car/{id}")
-    public ResponseEntity<?> delete(@PathVariable long id) {
-        final boolean deleted = carService.delete(id);
-        if (deleted) {
-            return new ResponseEntity<>(HttpStatus.OK);
+    @DeleteMapping(value = "/{employeeId}/driver/car/{id}")
+    public ResponseEntity<?> delete(@PathVariable long id,@PathVariable long employeeId) {
+        if (driverService.read(employeeId) != null) {
+            final boolean deleted = carService.delete(id);
+            if (deleted) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 }
